+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 
 ### App Center Distribute
 
-* **[Fix]** Fix a crash when downloading an update fails on Android 5+.
-* **[Improvement]** Improved error message logging when downloading an update fails on Android 5+.
+* **[Fix]** Fix a crash and improve logging when downloading an update fails on Android 5+.
 
 ___